django-template/requirements.txt
Django==5.0.4
djangorestframework==3.14.0
Pillow==10.2.0
django-cors-headers==4.3.1
python-dotenv==1.0.1
